+++ cassandra/src/main/scala/geotrellis/store/cassandra/package.scala
@@ -16,30 +16,22 @@
 
 package geotrellis.store
 
+import geotrellis.store.util._
 import cats.effect.Async
 import cats.syntax.functor._
 import com.datastax.oss.driver.api.core.CqlSession
 import com.datastax.oss.driver.api.core.cql.{AsyncResultSet, Statement}
 
 import java.math.BigInteger
-import java.util.concurrent.{CompletableFuture, CompletionStage}
 
 package object cassandra extends Serializable {
   implicit class BigIntOps(val self: BigInt) extends AnyVal {
     def asJava: BigInteger = new BigInteger(self.toByteArray)
   }
 
-  implicit class CompletableFutureOps[T](val self: CompletableFuture[T]) extends AnyVal {
-    def liftTo[F[_]: Async]: F[T] = Async[F].fromCompletableFuture(Async[F].delay(self))
-  }
-
-  implicit class CompletionStageOps[T](val self: CompletionStage[T]) extends AnyVal {
-    def liftTo[F[_]: Async]: F[T] = self.toCompletableFuture.liftTo
-  }
-
   implicit class CqlSessionOps(val self: CqlSession) extends AnyVal {
-    def closeF[F[_]: Async]: F[Unit] = self.closeAsync().liftTo.void
-    def executeF[F[_]: Async](statement: Statement[_]): F[AsyncResultSet] = self.executeAsync(statement).liftTo
+    def closeF[F[_]: Async]: F[Unit] = Async[F].liftCompletableFuture(self.closeAsync().toCompletableFuture).void
+    def executeF[F[_]: Async](statement: Statement[_]): F[AsyncResultSet] = Async[F].liftCompletableFuture(self.executeAsync(statement).toCompletableFuture)
   }
 
   implicit class AsyncResultSetOps(val self: AsyncResultSet) extends AnyVal {